+++ docs/conf.py
@@ -6,7 +6,7 @@ sys.path.insert(0, os.path.abspath(".."))
 project = "ravelpy"
 copyright = "2026, Derek Rowland"
 author = "Derek Rowland"
-release = "0.3.0"
+release = "0.3.1"
 
 extensions = [
     "myst_parser",